+++ src/app/cart/models/cart-item.model.ts
@@ -1,0 +1,5 @@
+import { Product } from './../../products/models/product.model';
+
+export class CartItem {
+  constructor(public product: Product, public price: number, public quant: number) {}
+}