+++ src/app/cart/services/cart.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Product } from './../../products/models/product.model';
+import { CartItem } from './../models/cart-item.model';
 
 @Injectable({
   providedIn: 'root'
@@ -10,7 +11,32 @@ export class CartService {
   getPurchasedProducts(): Array<Product> {
     return [
       new Product('Refrigerator1', 'NoFrost', 10000, true) ,
+      new Product('Refrigerator1', 'NoFrost', 10000, true) ,
       new Product('Refrigerator2', 'Frost', 15000, true)
     ];
+  }
+
+  getCartItems(): Array<CartItem> {
+    const products = this.getPurchasedProducts();
+    const cartItems = new Array<CartItem>();
+    products.forEach(element => {
+      this.pushProduct(cartItems, element);
+    });
+    return cartItems;
+  }
+
+  pushProduct(cartItems: Array<CartItem>, product: Product): Array<CartItem> {
+    let exist = false;
+    cartItems.forEach(item => {
+      if (item.product.name ===  product.name) {
+        item.price += product.price;
+        item.quant += 1;
+        exist = true;
+      }
+    });
+    if (!exist) {
+      cartItems.push(new CartItem(product, product.price, 1));
+    }
+    return cartItems;
   }
 }