+++ changelog.md
@@ -2,3 +2,11 @@
 2. Add Product Component
 3. Add ProductList component and ProductsService. Display Product List.
 4. Add Cart component and Cart Service. Display list of purchased products.
+
+5. Add products module, cart module, orders module and shared module 
+6. Use product component to display products in product list conponent
+7. Pass info about adding product to the basket from product compontn to product list via event emitter
+
+8. Add cart-item component
+9. Add cart-list component
+10. Add cart item delete and quantaty change functionality